+++ RasPISamples/scripts/weather.station/start.servers.sh
@@ -7,7 +7,7 @@ if [ "$STARTED" == "" ]
 then
   cd raspberry-pi4j-samples
   echo Starting multiplexer
-  cd NMEA.multiplexer
+  cd NMEA-multiplexer
   nohup ./mux.sh nmea.mux.weather.station.tcp.properties &
   echo Starting Snap loop
   cd ../RasPISamples